Alternative git editor (commented out) if running via ssh
If emacsserver is already running, logging in via ssh then creating a commit
seems to open a new emacs frame in the wrong X session. Uncomment the editor in
.gitconfig to fix it.

It might be possible to create a smarter script that looks if currently in
another TTY session, and not use emacsserver in that case?